+++ src/helpers/SettingsHelper.ts
@@ -1,4 +1,8 @@
-import { SETTING_PROJECTS } from './../constants/settings';
+import {
+  EXTENSION_NAME,
+  SETTING_CONFIG_DYNAMIC_FILE_PATH,
+  SETTING_PROJECTS
+} from './../constants/settings';
 import { parseWinPath } from './parseWinPath';
 import { Telemetry } from './Telemetry';
 import { Notifications } from './Notifications';
@@ -619,6 +623,40 @@ export class Settings {
       configFiles = configFiles.sort((a, b) => a.fsPath.localeCompare(b.fsPath));
       for await (const configFile of configFiles) {
         await Settings.processConfigFile(configFile);
+      }
+
+      // Check if there is a dynamic config file and use it to update the global config
+      const dynamicConfigPath =
+        Settings.globalConfig[`${CONFIG_KEY}.${SETTING_CONFIG_DYNAMIC_FILE_PATH}`];
+      if (dynamicConfigPath) {
+        try {
+          await window.withProgress(
+            {
+              location: vscode.ProgressLocation.Notification,
+              title: `${EXTENSION_NAME}: Reading dynamic config file...`
+            },
+            async () => {
+              const absFilePath = Folders.getAbsFilePath(dynamicConfigPath);
+              Logger.info(`Reading dynamic config file: ${absFilePath}`);
+              if (absFilePath) {
+                if (await existsAsync(absFilePath)) {
+                  const configFunction = require(absFilePath);
+                  const dynamicConfig = await configFunction(
+                    Object.assign({}, Settings.globalConfig)
+                  );
+
+                  if (dynamicConfig) {
+                    Settings.globalConfig = dynamicConfig;
+                    Logger.info(`Dynamic config file loaded`);
+                  }
+                }
+              }
+            }
+          );
+        } catch (e) {
+          Logger.error(`Error reading dynamic config file: ${dynamicConfigPath}`);
+          Logger.error((e as Error).message);
+        }
       }
     } catch (e) {
       Settings.globalConfig = undefined;